Piratha/dev

+++ src/main/java/com/ii/testautomation/controllers/TestTypesController.java
@@ -7,6 +7,7 @@ import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.response.common.ContentResponse;
 import com.ii.testautomation.response.common.FileResponse;
 import com.ii.testautomation.response.common.PaginatedContentResponse;
+import com.ii.testautomation.service.TestGroupingService;
 import com.ii.testautomation.service.TestTypesService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
@@ -21,9 +22,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @RestController
 @CrossOrigin
@@ -32,6 +31,8 @@ public class TestTypesController {
     private TestTypesService testTypesService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
+    @Autowired
+    private TestGroupingService testGroupingService;
 
     @PostMapping(EndpointURI.TEST_TYPE)
     public ResponseEntity<Object> insertTestTypes(@RequestBody TestTypesRequest testTypesRequest) {
@@ -76,6 +77,10 @@ public class TestTypesController {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getTestTypeNotExistCode(),
                     statusCodeBundle.getTestTypeIdNotFoundMessage()));
+        if (testGroupingService.existsByTestTypesId(id))
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getTestTypeDependentCode(),
+                    statusCodeBundle.getTestTypeDependentMessage()));
         testTypesService.deleteTestTypeById(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
@@ -117,6 +122,7 @@ public class TestTypesController {
 
         Map<String, List<Integer>> errorMessages = new HashMap<>();
         List<TestTypesRequest> testTypesRequestList;
+        Set<String> testTypeNames = new HashSet<>();
 
         try {
             if (multipartFile.getOriginalFilename().endsWith(".csv")) {
@@ -138,13 +144,17 @@ public class TestTypesController {
                 TestTypesRequest testTypesRequest = testTypesRequestList.get(rowIndex - 2);
 
                 if (!Utils.isNotNullAndEmpty(testTypesRequest.getName())) {
-                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameEmptyMessage(), rowIndex);
+                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getTestTypeNameEmptyMessage(), rowIndex);
+                } else if (testTypeNames.contains(testTypesRequest.getName())) {
+                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getTestTypeNameDuplicateMessage(), rowIndex);
+                } else {
+                    testTypeNames.add(testTypesRequest.getName());
                 }
                 if (!Utils.isNotNullAndEmpty(testTypesRequest.getDescription())) {
-                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectDescriptionEmptyMessage(), rowIndex);
+                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getTestTypeDescriptionEmptyMessage(), rowIndex);
                 }
                 if (testTypesService.isExistsTestTypeByName(testTypesRequest.getName())) {
-                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameAlReadyExistMessage(), rowIndex);
+                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getTestTypeNameAlReadyExistMessage(), rowIndex);
                 }
             }
